Bootloader update config (multicrate setup)
