+++ .cargo/config.toml
@@ -1,9 +1,4 @@
 [unstable]
-build-std-features = ["compiler-builtins-mem"]
-build-std = ["core", "compiler_builtins", "alloc"]
+bindeps = true
 
-[build]
-target = "x86_64-unknown-none"
 
-[target.'cfg(target_os = "none")']
-runner = "bootimage runner"
