+++ kernel/.cargo/config.toml
@@ -1,0 +1,9 @@
+[unstable]
+build-std-features = ["compiler-builtins-mem"]
+build-std = ["core", "compiler_builtins", "alloc"]
+bindeps = true
+
+[build]
+target = "x86_64-unknown-none"
+
+
