+++ kernel/Cargo.toml
@@ -1,0 +1,57 @@
+cargo-features = ["profile-rustflags"]
+
+[package]
+name = "rust-os"
+version = "0.1.0"
+edition = "2024"
+authors = ["Liam Storgaard"]
+
+
+[dependencies]
+acpi = "5.1.0"
+bootloader_api = "0.11.9"
+lazy_static = { version = "1.5.0", features = ["spin_no_std"] }
+noto-sans-mono-bitmap = "0.3.1"
+pc-keyboard = "0.8.0"
+pic8259 = "0.11.0"
+spin = "0.9.8"
+uart_16550 = "0.3.2"
+volatile = "0.2.6"
+x86_64 = "0.15.2"
+
+[dependencies.crossbeam-queue]
+version = "0.3.11"
+default-features = false
+features = ["alloc"]
+
+[dependencies.conquer-once]
+version = "0.2.0"
+default-features = false
+
+[dependencies.futures-util]
+version = "0.3.4"
+default-features = false
+features = ["alloc"]
+
+[package.metadata.bootimage]
+test-args = [
+    "-device", "isa-debug-exit,iobase=0xf4,iosize=0x04", "-serial", "stdio",
+    "-display", "none"
+]
+test-success-exit-code = 33
+test-timeout = 300
+
+[dependencies.bitvec]
+version = "1.0.1"
+default-features = false
+features = ["alloc"]
+
+[[test]]
+name = "should_panic"
+harness = false
+
+[[test]]
+name = "stack_overflow"
+harness = false
+
+
