+++ kernel/src/allocator.rs
