+++ kernel/src/allocator/alloc_info.rs
