+++ kernel/tests/heap_allocation.rs
