+++ kernel/tests/heap_deallocation.rs
